#### Problem

When adding a really high slot which skips many other slots, the code iterates through all slots from current next_slot all the way to the new slot whereas it should only ever need to iterate at most MAX_ENTRIES slots.

#### Summary of Changes

Detect this case and do a 0-fill, and then work in mod'ed space for the other fill.

Fixes #