@sakridge 2020-05-22
| code | entries: 2^10 (default) | entries: 2^10-1 | entries: 2^10, add ~2 * MAX_ENTRIES |
| -- | -- | -- | -- |
| new | 11.6ms |  11.4ms | 29.5ms |
| old | 12.0ms | 14.4ms | 0.072ms |

It's a 25% speedup with non-power-of-two. 5% with power-of-two, and at least 500x with adding >MAX_ENTRIES slot.